+++ MeshConverter/src/chunks/normals.h
@@ -22,12 +22,12 @@ struct NormalsChunk
 		SAFE_DELETE_ARRAY(normals);
 	}
 
-	size_t GetSize()
+	uint32_t GetSize()
 	{
 		if (count == 0)
 			return 0;
 
-		size_t size = sizeof(uint32_t);
+		uint32_t size = sizeof(uint32_t);
 		size += 
 			(sizeof(float) * 3)   // x, y, z
 			* count;
